example_common_library/pyproject.toml
[tool.poetry]
name = "Common_Library"
version = "1.3.8"
description = "A collection of methods used to access various Data Engineering services."
authors = ["Tim Bytnar <tbytnar@gmail.com>"]
readme = "README.md"
packages = [{include = "common_library"}]

[tool.poetry.dependencies]
python = "^3.10"
paramiko = "3.5.0"
pandas = "2.1.4"
Office365-REST-Python-Client = "2.5.13"
openpyxl = "3.1.3"
snowflake-connector-python = "3.12.4"
snowflake-sqlalchemy = "1.6.1"
SQLAlchemy = "1.4.54"
boto3 = "1.35.11"
s3fs = "2024.9.0"
python-semantic-release = "9.15.2"
pendulum = "3.0.0"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests/",
]
log_file_level = "info"
log_cli = 1

[tool.ruff]
line-length = 188
unfixable = ["F401"]

[tool.ruff.lint.pycodestyle]
max-line-length = 280

[tool.ruff.lint]
# Disable fix for unused imports (`F401`).
unfixable = ["F401"]


[tool.semantic_release]
version_variables = ["setup.py:__version__","common_library/__init__.py:__version__"]
version_toml = ["pyproject.toml:tool.poetry.version"]
major_on_zero = false
branch = "main"
upload_to_PyPI = false
upload_to_release = true
build_command = "sudo poetry build"

[tool.semantic_release.commit_parser_options]
allowed_types = [
    "build", "chore", "ci", "docs", "feat", "fix",
    "perf", "style", "refactor", "test"
]
minor_types = ["feat"]
patch_types = ["fix", "perf"]